Day 2 - auth and author

+++ controllers/todo.js
@@ -1,9 +1,10 @@
-const { Todo } = require('../models')
+const { Todo, User } = require('../models')
 
 class FancyTodo {
-    static add(req, res, next) {
+    static add(req, res) {
+        const { UserId } = req
         const { title, description, status, due_date } = req.body
-        Todo.create({ title, description, status, due_date })
+        Todo.create({ title, description, status, due_date, UserId })
             .then(data => {
                 res
                   .status(201)
@@ -24,23 +25,39 @@ class FancyTodo {
             })
     }   
 
-    static show(req, res) {
-        Todo.findAll()
-            .then(data => { 
-                res
-                  .status(200)
-                  .json({ data }) 
-            })
-            .catch(err => { 
-                res
-                  .status(500)
-                  .json({ msg: 'status code 500' }) 
-            })
+    static show(req, res, next) {
+        const { UserId } = req
+        if (UserId) {
+            Todo.findAll({ where: { UserId } })
+                .then(data => {
+                    if (data) {
+                        res
+                          .status(200)
+                          .json({ data }) 
+                    } else {
+                        // res
+                        //   .status(404)
+                        //   .json({ msg: 'NOT FOUND' })
+                        next(err)
+                    }
+                })
+                .catch(err => { 
+                    // res
+                    //   .status(500)
+                    //   .json({ msg: 'status code 500' })
+                    next(err)
+                })
+        } else {
+            res
+              .status(404)
+              .json({ msg: 'NOT FOUND' })
+        }
     }
 
     static show_id(req, res) {
+        const { UserId } = req
         const { id } = req.params
-        Todo.findByPk(id)
+        Todo.findOne({ where: { id, UserId } })
             .then(data => { 
                 if (data) {
                     res
@@ -55,13 +72,14 @@ class FancyTodo {
     }
 
     static delete(req, res) {
+        const { UserId } = req
         const { id } = req.params
-        Todo.destroy({ where: { id } })
-            .then(data => { 
+        Todo.destroy({ where: { id, UserId } })
+            .then(data => {
                 if (data) {
                     res
                       .status(204)
-                      .json({ data })
+                      .json({ msg: `success delete with id ${id}` })
                 } else {
                     res
                       .status(404)
@@ -71,19 +89,20 @@ class FancyTodo {
             .catch(err => { 
                 res
                   .status(500)
-                  .json({ msg: 'status code 500' })
+                  .json({ msg: err })
         })
     }
 
     static edit(req, res, next) {
+        const { UserId } = req
         const { id } = req.params
         const { title, description, status, due_date } = req.body
-        Todo.update({ title, description, status, due_date }, { where: { id } })
+        Todo.update({ title, description, status, due_date }, { where: { id, UserId } })
             .then(data => { 
                 if (data != 0) {
                     res
                     .status(200)
-                    .json({ data }) 
+                    .json({ data: req.body }) 
                 } else if (data == 0) {
                     res
                     .status(404)
@@ -96,12 +115,16 @@ class FancyTodo {
                     res
                       .status(400)
                       .json({ msg: `Invalid Input` })
-                }
+                } else if (err.name == 'jwt must be provided') {
+                    res
+                      .status(404)
+                      .json({ msg: 'NOT FOUND' })
+                } else {
                     res
                     .status(500)
                     .json({ err: err.message }) 
+                }
             })
-
     }
 }
